requirements.txt
Flask
gunicorn
numpy
pandas
scikit-learn
mysql.connector